address review: move cover UI to reachable comic editor
ComicPagesStage is unreachable in the running app — PipelineIssue
redirects /comicPages URLs to /comicScript where ComicScriptStage owns
the merged page editor. The cover UI initially placed in ComicPagesStage
never rendered for users. Moves the cover card into ComicScriptStage,
backed by the same stages.comicPages.cover record, and uses the
existing renderOpts so it picks up the user's image-gen drawer choice.

Also: extract-pages now seeds a blank cover.script from the LLM's
parsed `## Cover concept` section (preserves any user-edited cover);
drop unused getProfileLabels export; drop dead `comicPages` key from
VISUAL_STAGE_LABELS (stageId is never that value).
